Learning in Dreams-World Models On A Visualized Bin Packing Problem
The main contributions from this project are to measure the
impact of changing hyper-parameters from the world models
in a known case that is car racing, designing a visual representation
for the Bin Packing problem and also test how the
world models behave on the novel Bin Packing problem. These
serve as starting points towards continuing researching the
applicability of World Models for computer system problems.